Only check prebuild when PORTTESTING

+++ src/share/poudriere/common.sh
@@ -1081,18 +1081,20 @@ build_port() {
 		case ${phase} in
 		configure) [ -n "${PORTTESTING}" ] && markfs prebuild ${mnt} ;;
 		install-mtree)
-			mtree -X ${mnt}/poudriere/mtree.prebuildexclude -x \
-				-f ${mnt}/poudriere/mtree.prebuild \
-				-p ${mnt} > ${mnt}/tmp/preinst
-			if [ -s ${mnt}/tmp/preinst ]; then
-				msg "Filesystem touched before install:"
-				cat ${mnt}/tmp/preinst
+			if [ -n "${PORTTESTING}" ]; then
+				mtree -X ${mnt}/poudriere/mtree.prebuildexclude -x \
+					-f ${mnt}/poudriere/mtree.prebuild \
+					-p ${mnt} > ${mnt}/tmp/preinst
+				if [ -s ${mnt}/tmp/preinst ]; then
+					msg "Filesystem touched before install:"
+					cat ${mnt}/tmp/preinst
+					rm -f ${mnt}/tmp/preinst
+					bset ${MY_JOBID} status "preinst_fs_violation:${port}"
+					job_msg_verbose "Status for build ${port}: preinst_fs_violation"
+					return 1
+				fi
 				rm -f ${mnt}/tmp/preinst
-				bset ${MY_JOBID} status "preinst_fs_violation:${port}"
-				job_msg_verbose "Status for build ${port}: preinst_fs_violation"
-				return 1
 			fi
-			rm -f ${mnt}/tmp/preinst
 			;;
 		install) [ -n "${PORTTESTING}" ] && markfs preinst ${mnt} ;;
 		deinstall)
